% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Accidents map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Number of the state to extract the information}

\item{year}{Year from which represent the data}
}
\description{
Represent on a map the accidents occurs on a state and a year
}
\details{
This function requires libraries \code{dplyr},
\code{maps} and \code{graphics}
}
\examples{
data <- fars_map_state(state.num = 8, year = 2013)

}
\seealso{
\code{\link{make_filename}}

\code{\link{fars_read}}
}
