% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read years information}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{List of years from which to extract data}
}
\value{
A tible with the month and year read from the different csv files
  of the different years provided in \code{years}
}
\description{
Read month and year from one or several years passed in \code{year}
  input
}
\examples{
data <- fars_read_years(years = c(2013, 2014))

}
\seealso{
\code{\link{make_filename}}

\code{\link{fars_read}}
}
