% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize years information}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{List of years from which to extract data}
}
\value{
A tible with the number of accidents per month and year
  of the different years provided in \code{years} input
}
\description{
Summarize the information of the years  providing the number of
  accidents per month and year
}
\examples{
\dontrun{
data <- fars_summarize_years(years = c(2014, 2015))
}

}
\seealso{
\code{\link{fars_read_years}}
}
