% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create data file name including year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year from which to create the data file name. Year needs
to be a number or a character string representing a number, in
other case a warning will be raised and returned name will include
NA}
}
\value{
A character string of the file name include the year
}
\description{
Create data file name including year
}
\examples{
filename <- make_filename(year = 2013)

}
